+++ backend/Squirrel.Core/Squirrel.Core.DAL/Extensions/ApplicationBuilderExtensions.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
 using Squirrel.Core.DAL.Context;
 
 namespace Squirrel.Core.WebAPI.Extensions;
