+++ backend/Squirrel.Core/Squirrel.Core.WebAPI/Program.cs
@@ -1,4 +1,5 @@
 using Squirrel.Core.BLL.Extensions;
+using Squirrel.Core.DAL.Extensions;
 using Squirrel.Core.WebAPI.Extensions;
 using Squirrel.Core.WebAPI.Middlewares;
