+++ backend/Squirrel.SqlService/Squirrel.SqlService.WebApi/Program.cs
@@ -1,3 +1,6 @@
+using Squirrel.Core.DAL.Extensions;
+using Squirrel.Core.WebAPI.Extensions;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -5,6 +8,7 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSquirrelCoreContext(builder.Configuration);
 builder.Services.AddSwaggerGen();
 
 var app = builder.Build();
@@ -17,6 +21,8 @@ if (app.Environment.IsDevelopment())
 }
 
 app.UseHttpsRedirection();
+
+app.UseSquirrelCoreContext();
 
 app.UseAuthorization();
